test(crypto): split key-lifecycle.test.ts by lifecycle phase (crypto-mdzz)
Split key-lifecycle.test.ts (1021 LOC, 68 tests) into 4 focused files
plus shared fixtures. Test count preserved (68 -> 68). All files <500
LOC.

- key-lifecycle-unlock.test.ts (322 LOC, 23 tests)
- key-lifecycle-lock.test.ts (310 LOC, 20 tests)
- key-lifecycle-timers.test.ts (307 LOC, 19 tests)
- key-lifecycle-bucket-keys.test.ts (154 LOC, 6 tests)
- helpers/key-lifecycle-fixtures.ts (55 LOC)
